refactor: extract shared cycle-walking helper, simplify reductions
- Extract duplicated edges_to_cycle_order logic from HC→TSP and HC→BTSP
  into graph_helpers module (-110 lines)
- Use SimpleGraph::complete(k) in KClique→SubgraphIsomorphism
- Use .edges() directly in KClique→CBQ instead of O(n²) has_edge loop
- Replace assert! with graceful fallback in SAT→NAE-SAT extract_solution
